requirements.txt
backend==0.2.4.1
flasgger==0.9.7.1
Flask==3.1.0
Flask_Cors==5.0.0
Flask_JWT_Extended==4.7.1
flask_sqlalchemy==3.1.1
Werkzeug==3.1.3
